requirements.txt
blinker==1.8.2
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
flask==3.0.3
Flask-JWT-Extended==4.6.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.8.0
pymongo==4.7.2
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.1
werkzeug==3.0.3
zipp==3.19.0